separate e2e auth setup
