+++ packages/auth/package.json
@@ -17,6 +17,7 @@
   "exports": {
     ".": "./src/auth.ts",
     "./client": "./src/client.ts",
+    "./e2e": "./src/e2e.ts",
     "./hooks/auth-state": "./src/hooks/use-auth-state.ts",
     "./hooks/subscription": "./src/hooks/use-subscription.ts",
     "./next": "./src/next.ts",
